move determineDistribCodename to utils.ts
