+++ src/setup-ros-linux.ts
@@ -51,34 +51,7 @@ WE+F5FaIKwb72PL4rLi4
 `;
 
 // List of linux distributions that need http://packages.ros.org/ros/ubuntu APT repo
-const distrosRequiringRosUbuntu = [
-  "bionic",
-  "focal",
-];
-
-/**
- * Determines the Ubuntu distribution codename.
- *
- * This function directly source /etc/lsb-release instead of invoking
- * lsb-release as the package may not be installed.
- *
- * @returns Promise<string> Ubuntu distribution codename (e.g. "focal")
- */
- async function determineDistribCodename(): Promise<string> {
-	let distribCodename = "";
-	const options: im.ExecOptions = {};
-	options.listeners = {
-		stdout: (data: Buffer) => {
-			distribCodename += data.toString();
-		},
-	};
-	await utils.exec(
-		"bash",
-		["-c", 'source /etc/lsb-release ; echo -n "$DISTRIB_CODENAME"'],
-		options
-	);
-	return distribCodename;
-}
+const distrosRequiringRosUbuntu = ["bionic", "focal"];
 
 /**
  * Install ROS 2 on a Linux worker.
@@ -131,7 +104,7 @@ export async function runLinux() {
 	fs.writeFileSync(keyFilePath, openRoboticsAptPublicGpgKey);
 	await utils.exec("sudo", ["apt-key", "add", keyFilePath]);
 
-	const distribCodename = yield determineDistribCodename();
+	const distribCodename = await utils.determineDistribCodename();
 	if (distrosRequiringRosUbuntu.includes(distribCodename)) {
 		await utils.exec("sudo", [
 			"bash",
